+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 - Updated to use API v1.12.1
 - Corrected logic for checking for if PlayerSlots are full in `Split` ability.
 - Created `Flight of the Valkyrie` ability.
-  * When the creature is played, fill all open slots on your side of the field with 1/1 Norse Warriors
-  * Norse Warriors are defined as 1/1 normal cards. 
-- Added Resource image loading so that users no longer have to copy/paste the artwork folder. 
+  * When the creature is played, fill all open slots on your side of the field with 1/1 Norse Warriors.
+  * Norse Warriors are defined as 1/1 w/ Brittle.
+- Added Resource image loading so that users no longer have to copy/paste the artwork folder. Images are built with the .dll